requirements.txt
# Core
pydantic>=2.5,<3
pydantic-settings>=2.1,<3

# Database - PostgreSQL (cloud-portable via ABC)
sqlalchemy[asyncio]>=2.0,<3
psycopg2-binary>=2.9,<3

# Storage - S3 (behind ABC)
boto3>=1.34,<2

# PDF
pypdf>=4.0,<5

# AI
anthropic>=0.39,<1

# Dashboard
dash>=2.14,<3
dash-bootstrap-components>=1.5,<2
plotly>=5.18,<6
pandas>=2.0,<3

# CLI & report
python-dotenv>=1.0,<2
click>=8.1,<9
jinja2>=3.1,<4

# Testing
pytest>=7.0,<9

# Hashing / integrity
# hashlib is stdlib
